First Commit with gitignore
This is the first commit with mostly just changes to gitignore to only include files we need to edit or files associated with the creation and manipulation of the server itself. In order for this server to work right you would still need to include a valid install of Space Engineers Dedicated Server from Keen and the Torch server installation.